fixed compiler bug when using c++X0
git-svn-id: http://svn.boost.org/svn/boost/trunk@75068 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/uuid/seed_rng.hpp
@@ -139,7 +139,11 @@ private:
         }
 
         {
-            unsigned int rn[] = { std::rand(), std::rand(), std::rand() };
+            unsigned int rn[] = 
+                { static_cast<unsigned int>(std::rand())
+                , static_cast<unsigned int>(std::rand())
+                , static_cast<unsigned int>(std::rand())
+                };
             sha.process_bytes( (unsigned char const*)rn, sizeof( rn ) );
         }
